refactoring

+++ src/managers/ThreadManager.js
@@ -1,41 +1,46 @@
+const { constructor } = require('../data/battleMap');
 const { getChannel } = require('../util/getDiscordInfo');
+const BaseManager = require('./BaseManager');
 
-const threadManager = {
+class ThreadManager extends BaseManager {
 
-    setClient(client) {
-        this.client = client;
-    },
 
     /**
-     * creates a new thread in this channel
+     * set initial reference to instantiating client and interaction
+     * @param {Client|Interaction} param0 
+     */
+    constructor(client) {
+        super(client);
+    }
+
+    /**
+     * creates a new thread in this channel for PVE based battles
      * @param {String} user the the thread is created for
      * @returns {Thread} returns a reference to the thread so the messagehandler can send an init message
      */
-    async createThread(user, opponent) {
+    async createPveThread(battle) {
         if (!this.client) return;
-        const thread = (await (await getChannel(this.client, user.route)).threads.create({
-            name: `${user.username} vs. Level ${opponent.party[0].level} ${opponent.party[0].name}`
+        const thread = (await (await getChannel(this.client, battle.channel)).threads.create({
+            name: battle.name
         }));
         return thread.id;
-    },
+    }
 
     /**
      * deletes a thread in this channel
      * @param {String} user the user interacting with th
      */
-    async deleteThread(battle) {
+    async deleteThread(channel, threadName) {
         if (!this.client) return;
         // TODO: can we find a way to delete the message that point to this thread after the thread is deleted
-        (await getChannel(this.client, battle.player1.route)).threads.cache
+        (await getChannel(this.client, channel)).threads.cache
             .forEach(async(thread) => {
-                // TODO: fix the name references here
-                // if (thread.name === `${battle.p1.username} vs. Level ${battle.p2lead.level} ${battle.p2lead.name}`) {
-                //     await thread.delete();
-                // }
-                await thread.delete();
+                if (thread.name === threadName) {
+                    await thread.delete();
+                }
             })
     }
 
 }
 
-module.exports = threadManager;
+module.exports = ThreadManager;
